refactor docx.routes
